+++ domain_layer/src/main/java/com/nexaas/challenge/domain/model/ProductDomain.kt
@@ -3,8 +3,8 @@ package com.nexaas.challenge.domain.model
 import com.nexaas.challenge.domain.core.DomainModel
 
 data class ProductDomain (val name: String?,
-                          val quantity: Double?,
-                          val stock: Double?,
+                          val quantity: Int?,
+                          val stock: Int?,
                           val imageUrl: String?,
                           val price: Double?,
                           val tax: Double?,
